refactor: 테스트 코드가 reader 하위 패키지로 이동하도록 변경
